requirements.txt
asgiref==3.4.1
attrs==21.2.0
backports.entry-points-selectable==1.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
coverage==6.0.2
cryptography==35.0.0
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.3
dj-database-url==0.5.0
Django==3.2.8
django-cors-headers==3.10.0
django-extensions==3.1.3
django-heroku==0.3.1
django-oauth-toolkit==1.5.0
djangorestframework==3.12.4
drf-social-oauth2==1.0.9
factory-boy==3.2.0
Faker==9.5.0
filelock==3.3.1
gunicorn==20.1.0
idna==3.3
iniconfig==1.1.1
jwcrypto==1.0
oauthlib==3.1.1
packaging==21.0
platformdirs==2.4.0
pluggy==1.0.0
psycopg2==2.9.1
py==1.10.0
py-bcrypt==0.4
pycparser==2.20
PyJWT==2.3.0
pyparsing==2.4.7
pytest==6.2.5
pytest-django==4.4.0
python-dateutil==2.8.2
python-dotenv==0.19.1
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.7
virtualenv==20.8.1
whitenoise==5.3.0
wrapt==1.13.3
